Fix queries with multiple terms
Previously the query 'foo bar' would result in `q=text_t:foo+bar`.
That is, only the first term is using the correct field.  Meaning the
remaining terms have no effect on the results.  The fix is to set the
default field to `text_t` so all terms query against that field.